Add frame browser, phash dedup, and seconds-per-frame control
- Restore frame-level previewing with dual-handle range slider, thumbnail
  cards, step buttons, and transcript context for start/end frame selection
- Add upload-time phash dedup: removes similar frames immediately after
  extraction, configurable via Dedup Sensitivity in upload settings popover
- Replace FPS input with Seconds Per Frame (1-60), auto-computes fps and
  frames_per_minute for transcoding and analysis
- Add 4 API routes: /frames, /frames/<n>, /frames/<n>/thumb, /transcript
- Add _extract_frames() and _transcribe_video() post-transcode pipeline
- Worker uses pre-extracted frames with start_frame/end_frame range when
  available, falls back to extract_keyframes() with similarity_threshold
- Add imagehash dependency, update gpu_transcode fps param to float

+++ gpu_transcode.py
@@ -195,7 +195,7 @@ def build_transcode_command(
     output_path: str,
     width: int = 1280,
     height: int = 720,
-    fps: int = 1,
+    fps: float = 1,
     gpu_index: int = 0,
 ) -> List[str]:
     """
